fix(ci): file-line-limit — split adoption forge-gate tests into part_004
test_pr_monitor_adoption_part_002.py crossed the 1,500-line limit after this PR
added four forge-detection / legacy-snapshot adoption tests. Move them into a new
sibling part (test_pr_monitor_adoption_part_004.py) that imports the shared
adoption scaffolding (_MetadataFetcher, _metadata, _count) from part_002. No test
behavior changes; part_002 returns to 1,450 lines.
